requirements.txt
notebook
pandas
pdfplumber
amazon-textract-textractor
tabulate
ghostscript
numpy
pdf2image
Pillow
